computechange: Gotta be smart about modified ancestors.
Unfortunately, if we don't consider a version change in an ancestor, we
might miss a rename or file move that needs to happen.  I'm sure we
could do this more efficiently by precomputing rather than doing it from
the leaves of the tree up for each page, but oh well.